Mods to support machines with 64-bit addresses/pointers.  The changes are almost exclusively changing various ints to longs.  This includes changing int casts to long casts, and adding a few explicit long casts.
These changes appear to work for current DEC/OSF1 Alpha boxes using gcc.  Of
course we're screwed if we ever run into a machine regardless of address
size where sizeof(long) != sizeof(void*)...


git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/rdsslibs@6738 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ ui/ui_symbol.h
@@ -1,4 +1,4 @@
-/* $Id: ui_symbol.h,v 1.7 1998-02-26 21:18:46 burghart Exp $ */
+/* $Id: ui_symbol.h,v 1.8 1998-12-17 17:18:17 burghart Exp $ */
 /*
  * Global declarations for the symbol table module.
  */
@@ -23,7 +23,7 @@
  */
 typedef union usy_value
 {
-	int us_v_int;		/* Integer symbol value		*/
+	long us_v_int;		/* Integer symbol value		*/
 	float	us_v_float;	/* Floating point value		*/
 	date	us_v_date;	/* Date value			*/
 	char	*us_v_ptr;	/* Everything else		*/
